third_party/WebKit/LayoutTests/fast/dom/custom/imports-block-upgrade-in-import.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
var upgradeIds = [];
</script>
<x-a id="a1"></x-a>
<link rel="import" href="resources/import-block-upgrade-in-import-child.html">
<body>
<x-b id="b4"></x-b>
<x-a id="a5"></x-a>
<script>
test(function () {
    assert_array_equals(window.upgradeIds, ['a1', 'a2', 'a3', 'b0', 'b1', 'b2', 'a4', 'b3', 'b4', 'a5'], 
                        'Elements in the master and imports should be upgraded by a definition in an import.');
}, 'upgrade in the master');
</script>
</body>
